+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
- 
+
+## [2.3]
+
+- Sharding support added 
+- Moved permissions error message outside the embeds (clearer)
+- Updated command builder (use glob/path instead of fs)
+- HZV revamp (more spacing)
+  
+
 ## [2.2.0] 
  
 - `+toggle` for admins who want to disable a command or category in their server